+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - none
 
 # v2022.6.1-beta
+- error - add sqlite errcode SQLITE_ERROR_ZSQL_NULL
 - jsbaton - migrate function dbTableInsert() to new unified jsbaton
 - jsbaton - migrate functions dbClose(), dbMemoryLoadOrSave(), dbNoop(), dbOpen() to new, streamlined jsbaton2
 - add js-function jsbatonPushValue() to dynamically push value to jsbaton